standardize placement of desktop nav, rename mobileheader -> main-header
