+++ src/component/container/Home.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 
 // components
 import NavigatorMobile from '../navigation/NavigatorMobile'
-import MobileHeader from '../partial/MobileHeader'
+import MainHeader from '../partial/MainHeader'
 import PostViewer from './PostViewer'
 
 // style
@@ -14,7 +14,7 @@ class Home extends React.Component {
         if(!this.props.posts || !this.props.self) return (null)
         return (
             <div className='Home'>
-                <MobileHeader title={'Friends'}/>
+                <MainHeader title={'Friends'}/>
                 <PostViewer includeUsers={this.props.self.friends.concat([this.props.self._id])} view={'LIST'} />
                 <NavigatorMobile />
             </div>
