+++ src/css/HeaderNav.css
@@ -1,0 +1,33 @@
+.page-header_nav {
+    flex: 1;
+}
+
+.page-header_nav-list-container{
+    display: none;
+}
+
+@media (min-width: 700px) {
+    .page-header_nav{
+        justify-content: flex-end;
+        display: flex;
+        flex: 1;
+    }
+
+    .page-header_nav-list-container {
+        width: 90%;
+        margin-left: 10%;
+        display: flex;
+        justify-content: flex-end;
+    }
+
+    .page-header_nav-list {
+        width: 100%;
+        display: flex;
+        flex-direction: row;
+        justify-content: flex-end;
+    }
+
+    .page-header_nav-list > a {
+        margin-left: 1em;
+    }
+}
