--- src/css/MobileHeader.css
@@ -1,80 +1,0 @@
-.page-header {
-    background-color: white;
-    border-bottom: 1px solid #efefef;
-    margin-bottom: 2rem;
-}
-
-.page-header_container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 90%;
-    margin: 0 auto;
-}
-
-.page-header_logo {
-    display: flex;
-    flex: 1;
-    justify-content: flex-start;
-    box-sizing: border-box;
-    width: 25%;
-}
-
-.page-header_title {
-    color: tomato;
-}
-
-.page-header_logout-link {
-    display: flex;
-    justify-content: flex-end;
-}
-
-.page-header_nav {
-    flex: 1;
-}
-
-.page-header_nav-list-container{
-    display: none;
-}
-
-@media (min-width: 700px) {
-    .page-header_nav-list-container{
-        display: flex;
-    }
-
-    .page-header {
-        font-size: 1.4rem;
-    }
-    .page-header_logo {
-        flex: 1;
-    }
-
-    .page-header_container {
-        width: 70%;
-        margin: 0 auto;
-    }
-
-    .page-header_logo {
-        justify-content: flex-start;
-    }
-
-    .page-header_nav{
-        justify-content: flex-end;
-        display: flex;
-        flex: 1;
-    }
-
-    .page-header_nav-list-container {
-        width: 90%;
-        margin-left: 10%;
-        display: flex;
-        justify-content: flex-end;
-    }
-
-    .page-header_nav-list {
-        width: 100%;
-        display: flex;
-        flex-direction: row;
-        justify-content: space-around;
-    }
-}
